refactor: FriendProfileView MVVM 분리 (ViewModel 로직 분리), 친구 요청 상태 서버 확인 추가
